Add CharacterSelectionManager
