+++ common/src/main/java/com/wynntils/gui/widgets/ClassSelectionEditButton.java
@@ -9,7 +9,7 @@ import com.wynntils.gui.render.FontRenderer;
 import com.wynntils.gui.render.RenderUtils;
 import com.wynntils.gui.render.Texture;
 import com.wynntils.gui.screens.CharacterSelectorScreen;
-import com.wynntils.wynn.utils.ContainerUtils;
+import com.wynntils.wynn.model.CharacterSelectionManager;
 import java.util.List;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.components.AbstractButton;
@@ -18,14 +18,12 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.world.inventory.AbstractContainerMenu;
-import org.lwjgl.glfw.GLFW;
 
 public class ClassSelectionEditButton extends AbstractButton {
     private static final List<Component> TOOLTIP = List.of(
             new TranslatableComponent("screens.wynntils.characterSelection.edit.name").withStyle(ChatFormatting.YELLOW),
             new TranslatableComponent("screens.wynntils.characterSelection.edit.discussion")
                     .withStyle(ChatFormatting.GRAY));
-    private static final int EDIT_BUTTON_SLOT = 8;
     private final CharacterSelectorScreen characterSelectorScreen;
 
     public ClassSelectionEditButton(
@@ -39,7 +37,7 @@ public class ClassSelectionEditButton extends AbstractButton {
         AbstractContainerMenu menu =
                 characterSelectorScreen.getActualClassSelectionScreen().getMenu();
 
-        ContainerUtils.clickOnSlot(EDIT_BUTTON_SLOT, menu.containerId, GLFW.GLFW_MOUSE_BUTTON_LEFT, menu.getItems());
+        CharacterSelectionManager.editCharacters(menu);
     }
 
     @Override
