+++ common/src/main/java/com/wynntils/gui/widgets/PlayButton.java
@@ -9,7 +9,7 @@ import com.wynntils.gui.render.FontRenderer;
 import com.wynntils.gui.render.RenderUtils;
 import com.wynntils.gui.render.Texture;
 import com.wynntils.gui.screens.CharacterSelectorScreen;
-import com.wynntils.wynn.utils.ContainerUtils;
+import com.wynntils.wynn.model.CharacterSelectionManager;
 import java.util.List;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.components.AbstractButton;
@@ -18,7 +18,6 @@ import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
-import org.lwjgl.glfw.GLFW;
 
 public class PlayButton extends AbstractButton {
     private static final List<Component> TOOLTIP = List.of(
@@ -37,14 +36,7 @@ public class PlayButton extends AbstractButton {
     public void onPress() {
         if (characterSelectorScreen.getSelected() != null) {
             int slot = characterSelectorScreen.getSelected().getClassInfo().slot();
-            ContainerUtils.clickOnSlot(
-                    slot,
-                    characterSelectorScreen.getActualClassSelectionScreen().getMenu().containerId,
-                    GLFW.GLFW_MOUSE_BUTTON_LEFT,
-                    characterSelectorScreen
-                            .getActualClassSelectionScreen()
-                            .getMenu()
-                            .getItems());
+            CharacterSelectionManager.playWithCharacter(slot);
         }
     }
